rhythm-party: fix browser sync (float ping) + ASCII-safe UI
The browser sends ping timestamps as floats (sub-millisecond wall clock),
but the conductor decoded `c` as int64, so Go's JSON unmarshal failed and
closed the socket (1007) right after the handshake -- every web client
fell back to "solo (offline)". Decode `c` as float64 (and echo it back as
one). My Go smoke client sent clean integers, which is why it never hit it.

Also: the default Godot font has no emoji/symbol glyphs (they render as
tofu boxes), so make all on-screen text ASCII-safe, and simplify the
same-origin WebSocket URL lookup to reference JavaScriptBridge directly.

+++ rhythm-party/conductor/main.go
@@ -131,14 +131,14 @@ func (h *hub) tickPresence(every time.Duration) {
 // --- messages ---------------------------------------------------------------
 
 type clientMsg struct {
-	T string `json:"t"`           // "ping" | "hit"
-	C int64  `json:"c,omitempty"` // client send time (ms) for ping round-trip
+	T string  `json:"t"`           // "ping" | "hit"
+	C float64 `json:"c,omitempty"` // client send time (ms); float — JS clocks are sub-ms
 }
 
 type pongMsg struct {
-	T string `json:"t"` // "pong"
-	C int64  `json:"c"` // echoed client time
-	S int64  `json:"s"` // server now (ms)
+	T string  `json:"t"` // "pong"
+	C float64 `json:"c"` // echoed client time
+	S int64   `json:"s"` // server now (ms)
 }
 
 type syncMsg struct {
